libraries/sql.lib.php file added. Functions in the sql.php transfered into sql.lib.php
